@@ -3,12 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/core": "^10.0.10",
-    "node-sass": "^4.11.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-icons": "^3.6.1",
-    "react-router-dom": "^5.0.0",
+    "@emotion/core": "^10.0.28",
+    "node-sass": "^4.14.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-icons": "^3.10.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "2.1.8"
   },
   "scripts": {
@@ -29,7 +29,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "bootstrap": "^4.3.1",
-    "reactstrap": "^8.0.0"
+    "bootstrap": "^4.5.0",
+    "reactstrap": "^8.4.1"
   }
 }
