@@ -21,8 +21,10 @@ const AddGroceryItemForm = () => {
             <h2 className={styles.header}>ADD NEW ITEM</h2>
             <form
               className={styles.form}
-              onSubmit={(event) =>
-                context.addItem(event, name, quantity, category)
+              onSubmit={(event) =>{
+                context.addItem(event, {name, quantity, category})
+              }
+               
               }>
               <Input
                 onChange={(e) => setName(e.target.value)}
@@ -52,7 +54,7 @@ const AddGroceryItemForm = () => {
   );
 };
 
-const AddPriceItemForm = ({ context, handleInputChange, state }) => {
+const AddPriceItemForm = ({ name }) => {
   const [price, setPrice] = useState('');
   const [buyer, setBuyer] = useState('');
 
@@ -64,7 +66,7 @@ const AddPriceItemForm = ({ context, handleInputChange, state }) => {
             <h2 className={styles.header}>ADD PRICE OF ONE ITEM</h2>
             <form
               className={styles.form}
-              onSubmit={(event) => context.updateItem(event, state)}>
+              onSubmit={(event) => context.updateItem(event, { price, buyer, name })}>
               <Input
                 onChange={(e) => setPrice(e.target.value)}
                 value={price}
@@ -92,13 +94,13 @@ const AddPriceItemForm = ({ context, handleInputChange, state }) => {
   );
 };
 
-const Form = ({ formType }) => {
+const Form = ({ formType, name }) => {
   if (formType === FormTypes.addItem) {
     return <AddGroceryItemForm />;
   }
 
   if (formType === FormTypes.addPrice) {
-    return <AddPriceItemForm />;
+    return <AddPriceItemForm name={name} />;
   }
 };
 
