@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Root from './views/Root/Root';
+import Root from './views/Root/Root.js';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 ReactDOM.render(<Root />, document.getElementById('root'));
